Adding a second controller
Had to do some timing voodoo...not terribly happy with it,
but it seems to work.

Root issue is that controllers send back their responses
at slightly different start times, so we have to manage to
poll them when they're both in their data phase.

This could be obivated with serial requests, but that's no fun.